[APG-2119] add domain event service
